Fix mic icon color for radio
Previously the mic icon would be white when not speaking, yellow when speaking in proximity, and red when speaking in prox or over radio.

now white when not speaking, yellow when speaking in prox, and red when speaking over the radio.

Small change, but this has bugged me for months lmao
Seems to be the intended behavior anyway